requirements.txt
fastapi
uvicorn
pyelftools
capstone
python-multipart